(fix) merge unit test for BytesUtil
